Fix circular dependencies in `task` package
